requirements.txt
 
streamlit==1.28.1

 
google-generativeai==0.3.2
openai==1.3.7

 
opencv-python-headless==4.8.1.78
pillow==10.1.0
imageio-ffmpeg==0.4.9
faster-whisper==1.0.3

 
pandas==2.1.4
numpy==1.26.4

streamlit-option-menu==0.3.6
streamlit-lottie==0.0.5